Refactor orbit simulation to separate module
Moved orbit simulation logic from gui.py to a new orbit_simulator.py module. Removed related imports and UI elements from gui.py to improve code organization and separation of concerns.
